backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
email-validator>=2.1.0
bcrypt>=4.2.0
pyjwt>=2.9.0
pywebpush>=2.0.0
apscheduler>=3.10.0
python-multipart>=0.0.12
